Move gql_api_builder to global config for reuse
