+++ reconcile/test/test_typed_queries/conftest.py
@@ -2,15 +2,9 @@ from collections.abc import (
     Callable,
     Mapping,
 )
-from typing import (
-    Any,
-    Optional,
-)
-from unittest.mock import create_autospec
+from typing import Any
 
 import pytest
-
-from reconcile.utils.gql import GqlApi
 
 
 @pytest.fixture
@@ -20,15 +14,5 @@ def query_func() -> Callable[[Mapping], Callable]:
             return data
 
         return query_func
-
-    return builder
-
-
-@pytest.fixture
-def gql_api_builder() -> Callable[[Optional[Mapping]], GqlApi]:
-    def builder(data: Optional[Mapping] = None) -> GqlApi:
-        gql_api = create_autospec(GqlApi)
-        gql_api.query.return_value = data
-        return gql_api
 
     return builder
